Move IMDb pydantic models to common file
