--- art_graph/cinema_data_providers/imbd_non_commercial/title_crew.py
@@ -1,8 +1,0 @@
-from pydantic import BaseModel
-from typing import List, Optional
-
-
-class TitleCrew(BaseModel):
-    tconst: str
-    directors: Optional[List[str]] = None
-    writers: Optional[List[str]] = None
