--- art_graph/cinema_data_providers/imbd_non_commercial/title_episode.py
@@ -1,9 +1,0 @@
-from pydantic import BaseModel
-from typing import Optional
-
-
-class TitleEpisode(BaseModel):
-    tconst: str
-    parentTconst: str
-    seasonNumber: Optional[int] = None
-    episodeNumber: Optional[int] = None
