--- art_graph/cinema_data_providers/imbd_non_commercial/title_principals.py
@@ -1,11 +1,0 @@
-from pydantic import BaseModel
-from typing import Optional
-
-
-class TitlePrincipals(BaseModel):
-    tconst: str
-    ordering: int
-    nconst: str
-    category: str
-    job: Optional[str] = None
-    characters: Optional[str] = None
