--- art_graph/cinema_data_providers/imbd_non_commercial/title_ratings.py
@@ -1,7 +1,0 @@
-from pydantic import BaseModel
-
-
-class TitleRatings(BaseModel):
-    tconst: str
-    averageRating: float
-    numVotes: int
